libs_snap/linearanomaly/requirements.txt
jupyter
traitlets>=5.1.1
pygments>=2.4.1
awkward>=1.10.1
typing-extensions>=4.5.0

# pytorch / deep learning
torchmetrics==0.11.4
wandb>=0.16.0
POT>=0.9.3

# Parameterised fitting
zfit>=0.20.3

# hydra and omegaconf
hydra-core==1.3.2
hydra-colorlog==1.2.0
omegaconf>=2.2.3

# plotting
matplotlib==3.7.1
seaborn==0.12.2
plotly==5.13.1

# data science
numpy==1.23.5
scipy==1.10.1
scikit-learn==1.2.2
scikit-image>=0.22.0
scikit-hep==5.0.0
pandas==1.5.3
tables==3.8.0
h5py>=3.10.0
pykeops==2.1.1

# other
black==23.1.0
pylint==2.17.0
dotmap==1.3.30
pyrootutils==1.0.4
pre-commit==3.2.0
rich==13.3.2
pytest==7.2.2
pytest-cov==4.1.0
